# src/main/resources/static/swagger-api/swagger.yaml
...
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "st-schools"
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /api/v1/admin/user/{id}:
    get:
      tags:
        - "AdminController"
      summary: "Get user by id"
      operationId: "getUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
      security:
        - token: []
  /api/v1/admin/user/all:
    get:
      tags:
        - "AdminController"
      summary: "Get users from database"
      operationId: "getAllUsers"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
      security:
        - token: []
  /api/v1/admin/graphql/user:
    post:
      tags:
        - "AdminController"
      summary: "Get user by GraphQL query"
      operationId: "getUserByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ user(id: 2) { id email firstName lastName address phone  birthday} }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  email:
                    type: "string"
                  firstName:
                    type: "string"
                  lastName:
                    type: "string"
                  address:
                    type: "string"
                  phone:
                    type: "string"
                  birthday:
                    type: "string"
                example:
                  id: 2
                  email: "thongchuthanh2000@gmail.com"
                  firstName: "Cheng"
                  lastName: "Tang Yu"
                  address: "BR-VT"
                  phone: "0918948074"
                  birthday: "12/11/2000"
      security:
        - token: []
  /api/v1/admin/graphql/user/all:
    post:
      tags:
        - "AdminController"
      summary: "Get all users by GraphQL query"
      operationId: "getAllUsersByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ user(id: 2) { id email firstName lastName address phone  birthday} }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    email:
                      type: "string"
                    firstName:
                      type: "string"
                    lastName:
                      type: "string"
                    address:
                      type: "string"
                    phone:
                      type: "string"
                    birthday:
                      type: "string"
                    example:
                      id: 2
                      email: "thongchuthanh2000@gmail.com"
                      firstName: "Cheng"
                      lastName: "Tang Yu"
                      address: "BR-VT"
                      phone: "0918948074"
                      birthday: "12/11/2000"

      security:
        - token: []

#  /api/v1/admin/graphql/orders:
#    post:
#      tags:
#        - "AdminController"
#      summary: "Get all orders by GraphQL query"
#      operationId: "getAllOrdersQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ orders { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/OrderResponse"
#      security:
#        - token: []
#
#  /api/v1/admin/graphql/order:
#    post:
#      tags:
#        - "AdminController"
#      summary: "Get orders by user Email GraphQL query"
#      operationId: "getUserOrdersByEmailQuery"
#      requestBody:
#        required: true
#        content:
#          'application/json':
#            schema:
#              type: object
#              properties:
#                query:
#                  type: "string"
#                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
#      responses:
#        "200":
#          description: "OK"
#          content:
#            'application/json':
#              schema:
#                type: array
#                items:
#                  $ref: "#/components/schemas/OrderResponse"
#      security:
#        - token: []
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        email:
          type: "string"
          example: "thongchuthanh2000@test.com"
        firstName:
          type: "string"
          example: "Cheng"
        lastName:
          type: "string"
          example: "Tang Yu"
        avatar:
          type: "string"
          example: ""
        address:
          type: "string"
          example: "BR-VT"
        phone:
          type: "string"
          example: "0918948074"
        provider:
          type: "string"
          example: "LOCAL"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
...